+++ ReleaseNotes.md
@@ -2,6 +2,7 @@
 
 ## 2.6.1
 
+* Fixed range-valued `[validate]` parsing
 * Fixed an error message for invalid `[validate]` use on collections
 
 ## 2.6.0